requirements.txt
Django>=5.0,<6.0
celery>=5.4
redis>=5.0
python-dotenv>=1.0
